Redone the refactoring for most components, based on the current copy of the proxy. TODO: fitting and undo/redo methods
